Move tweet-id-key.ts to lib/storage/
