+++ src/content-twitter/store.ts
@@ -1,7 +1,7 @@
 import { PayloadAction, configureStore, createSlice } from '@reduxjs/toolkit';
 import { Middleware } from 'redux';
 import { createLogger } from 'redux-logger';
-import { TweetIDKey, toTweetIDKey } from '~/lib/tweet-id-key';
+import { TweetIDKey, toTweetIDKey } from '~/lib/storage/tweet-id-key';
 import { TweetID } from '~/lib/tweet/tweet';
 
 export interface ButtonStateNone {
