+++ src/lib/storage/tweet-id-key.ts
@@ -1,4 +1,4 @@
-import { TweetID } from './tweet/tweet';
+import { TweetID } from '../tweet/tweet';
 
 export type TweetIDKey = `tweet_${TweetID}`;
