+++ src/lib/storage/tweet.ts
@@ -3,8 +3,8 @@ import { tweetJSONSchema, tweetsJSONSchema } from '~/jsonschema/tweet';
 import { JSONSchemaValidationError } from '~/validate-json/error';
 import validateTweet from '~/validate-json/validate-tweet';
 import validateTweets from '~/validate-json/validate-tweets';
-import { isTweetIDKey, toTweetID, toTweetIDKey } from '../tweet-id-key';
 import { Tweet, TweetID } from '../tweet/tweet';
+import { isTweetIDKey, toTweetID, toTweetIDKey } from './tweet-id-key';
 
 export const saveTweets = async (tweets: Tweet[]): Promise<void> => {
   // JSON Schema validation
